requirements.txt
# Django and Web Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9

# Machine Learning - Computer Vision
opencv-python==4.9.0.80
deepface==0.0.90
tf-keras==2.16.0
pillow==10.2.0

# Machine Learning - Audio/Speech  
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# NLP and AI
openai==1.7.2
transformers==4.36.2

# Data Processing
pandas==2.1.4
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
